+++ src/helix-ui/elements/hx-tab/index.js
@@ -1,4 +1,4 @@
-import { HXElement } from './HXElement';
+import { HXElement } from '../HXElement.js';
 
 /**
  * Fires when non-current tab is clicked.
