+++ src/helix-ui/elements/index.js
@@ -32,11 +32,11 @@ export { HXSearchAssistanceElement } from './hx-search-assistance/index.js';
 export { HXSearchElement } from './HXSearchElement';
 export { HXSelectControlElement } from './HXSelectControlElement';
 export { HXSelectElement } from './HXSelectElement';
-export { HXTabElement } from './HXTabElement';
-export { HXTabcontentElement } from './HXTabcontentElement';
-export { HXTablistElement } from './HXTablistElement';
-export { HXTabpanelElement } from './HXTabpanelElement';
-export { HXTabsetElement } from './HXTabsetElement';
+export { HXTabElement } from './hx-tab/index.js';
+export { HXTabcontentElement } from './hx-tabcontent/index.js';
+export { HXTablistElement } from './hx-tablist/index.js';
+export { HXTabpanelElement } from './hx-tabpanel/index.js';
+export { HXTabsetElement } from './hx-tabset/index.js';
 export { HXTextControlElement } from './HXTextControlElement';
 export { HXTextareaControlElement } from './HXTextareaControlElement';
 export { HXToastElement } from './HXToastElement';
